Deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
numpy
scikit-learn==1.5.1
Pillow
plotly